+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Fix USB kernel panics on iMac7,1
 - Fix AppleALC support in Mojave
 - Fix TeraScale 1 GPU detection
+- Enable Graphics Acceleration on legacy GPUs by default
 
 ## 0.1.0
 - Fix crash on iMacs with Metal GPUs